requirements.txt
asgiref==3.2.7
Django==3.0.6
django-taggit==1.2.0
django-tinymce4-lite==1.8.0
jsmin==2.2.2
mysqlclient==1.4.6
Pillow==7.1.2
psycopg2==2.8.5
pytz==2020.1
sqlparse==0.3.1
